+++ src/pages/savings/recommendations/RecommendSavings.vue
@@ -1,10 +1,25 @@
 <script setup>
-import { ref } from 'vue';
+import { ref, watch } from 'vue';
 import SavingMyProductSlider from '@/components/savings/SavingMyProductSlider.vue';
 import BankbookProductList from '@/components/bankbook/BankbookProductList.vue';
 import RecommendationLayout from '@/components/layouts/RecommendationLayout.vue';
-// ✅ 이 부분 추가!
+import savingApi from '@/api/savings';
+
 const selectedId = ref(null);
+const savings = ref([]);
+
+// ✅ selectedId 자체를 감시하세요.
+watch(
+  selectedId,
+  async (newId) => {
+    if (newId != null) {
+      const data = await savingApi.getRecommendSavings(newId);
+      savings.value = data;
+      console.log(savings.value);
+    }
+  },
+  { immediate: true }
+);
 </script>
 
 <template>
@@ -13,10 +28,21 @@ const selectedId = ref(null);
       <SavingMyProductSlider @select="(id) => (selectedId = id)" />
     </template>
 
-    <template #highlight> 지금보다 더 좋은 적금 상품을 찾았어요! </template>
+    <!-- 다시 불러오기 의 경우 -->
+    <template #highlight v-if="selectedId == -1">
+      혹시 이런 상품은 어떠세요?
+    </template>
+    <!-- 추천하는 적금이 있으면 -->
+    <template #highlight v-else-if="savings.length > 0">
+      지금보다 더 좋은 적금 상품을 찾았어요!
+    </template>
 
-    <template #content>
-      <BankbookProductList :selectedId="selectedId" />
+    <!-- 더 좋은 적금이 없으면 -->
+    <template #highlight v-else>
+      지금보다 더 좋은 상품을 찾지 못했어요.
+    </template>
+    <template #content v-if="savings.length > 0">
+      <BankbookProductList :selectedId="selectedId" :savings="savings" />
     </template>
   </RecommendationLayout>
 </template>
